pyproject.toml
[tool.poetry]
name = "bloodsugar"
version = "0.1.0"
description = ""
authors = ["이주환 <118493627+LeeJuHwan@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.14.0"
django = "^4.2.6"
djangorestframework-simplejwt = "^5.3.0"
python-dotenv = "^1.0.0"
drf-spectacular = "^0.26.5"
dj-rest-auth = "^5.0.1"
django-allauth = "^0.58.2"
django-cors-headers = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
